Add files via upload
Node and npm are installed. Here are the versions I'll be using while making this tutorial:

$ node --version
v16.13.2

$ npm --version
8.1.2
Installing npm adds two commands to the system—npm and npx—both of which I'll be using while making this tutorial.

Git is installed. Here's the version I'll be using while making this tutorial:

$ git --version
git version 2.29.1.windows.1